refactor. Changed localTime widget into a web component.
