+++ src/blog.html
@@ -24,12 +24,7 @@
     </main>
     <section class="sidebar-right">
       <hr />
-      <section class="widget">
-        <header>
-          <span>My Local Time</span>
-        </header>
-        <time id="localTimeDisplay" datetime="2025-01-01T00:00:00-0500">00:00:00, January 1, 2025 </time>
-      </section>
+      <local-time-widget></local-time-widget>
       <hr />
     </section>
     <script type="module" src="./scripts/main.ts"></script>
