+++ src/scripts/components/localTime/localTime.html
@@ -1,0 +1,6 @@
+<section class="widget">
+  <header>
+    <span>My Local Time</span>
+  </header>
+  <time id="localTimeDisplay">00:00:00</time>
+</section>
