+++ src/scripts/main.ts
@@ -1,11 +1,6 @@
 import '../styles/style.css'
 
-import './components/siteNav'
-
-import { initializeLocalTime } from './widgets/localTime'
-
-document.addEventListener('DOMContentLoaded', () => {
-    initializeLocalTime();
-})
+import './components/siteNav/siteNav'
+import './components/localTime/localTime'
 
 console.log('Well, aren\'t you a curious one? 😉')
